+++ __tests__/index.test.js
@@ -360,6 +360,9 @@ describe('ServerlessVpcPlugin', () => {
           ['Public', '10.0.40.0/22'],
           ['DB', '10.0.44.0/22']]),
         ],
+        ['App', ['10.0.0.0/21', '10.0.16.0/21', '10.0.32.0/21']],
+        ['Public', ['10.0.8.0/22', '10.0.24.0/22', '10.0.40.0/22']],
+        ['DB', ['10.0.12.0/22', '10.0.28.0/22', '10.0.44.0/22']],
       ];
 
       const expected = new Map(parts);
@@ -1390,503 +1393,6 @@ describe('ServerlessVpcPlugin', () => {
       const actual = plugin.buildLambdaVPCEndpointSecurityGroup({
         name: 'MyLambdaEndpointSecurityGroup',
       });
-      expect(actual).toEqual(expected);
-    });
-  });
-
-  describe('#buildNetworkAcl', () => {
-    it('builds a Network ACL', () => {
-      const expected = {
-        AppNetworkAcl: {
-          Type: 'AWS::EC2::NetworkAcl',
-          Properties: {
-            Tags: [
-              {
-                Key: 'STAGE',
-                Value: 'dev',
-              },
-              {
-                Key: 'Name',
-                Value: {
-                  'Fn::Join': [
-                    '-',
-                    [
-                      {
-                        Ref: 'AWS::StackName',
-                      },
-                      'app',
-                    ],
-                  ],
-                },
-              },
-            ],
-            VpcId: {
-              Ref: 'VPC',
-            },
-          },
-        },
-      };
-      const actual = plugin.buildNetworkAcl('App');
-      expect(actual).toEqual(expected);
-    });
-  });
-
-  describe('#buildNetworkAclEntry', () => {
-    it('builds an ingress Network ACL entry', () => {
-      const expected = {
-        PublicNetworkAclIngress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '0.0.0.0/0',
-            NetworkAclId: {
-              Ref: 'PublicNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-      };
-      const actual = ServerlessVpcPlugin.buildNetworkAclEntry(
-        'PublicNetworkAcl', '0.0.0.0/0',
-      );
-      expect(actual).toEqual(expected);
-    });
-
-    it('builds an egress Network ACL entry', () => {
-      const expected = {
-        PublicNetworkAclEgress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '0.0.0.0/0',
-            NetworkAclId: {
-              Ref: 'PublicNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-      };
-      const actual = ServerlessVpcPlugin.buildNetworkAclEntry(
-        'PublicNetworkAcl', '0.0.0.0/0', { Egress: true },
-      );
-      expect(actual).toEqual(expected);
-    });
-  });
-
-  describe('#buildNetworkAclAssociation', () => {
-    it('builds a network ACL association', () => {
-      const expected = {
-        AppSubnetNetworkAclAssociation1: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'AppSubnet1',
-            },
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-          },
-        },
-      };
-      const actual = ServerlessVpcPlugin.buildNetworkAclAssociation('App', 1);
-      expect(actual).toEqual(expected);
-    });
-  });
-
-  describe('#buildPublicNetworkAcl', () => {
-    it('builds the public network ACL', () => {
-      const expected = {
-        PublicNetworkAcl: {
-          Type: 'AWS::EC2::NetworkAcl',
-          Properties: {
-            Tags: [
-              {
-                Key: 'STAGE',
-                Value: 'dev',
-              },
-              {
-                Key: 'Name',
-                Value: {
-                  'Fn::Join': [
-                    '-',
-                    [
-                      {
-                        Ref: 'AWS::StackName',
-                      },
-                      'public',
-                    ],
-                  ],
-                },
-              },
-            ],
-            VpcId: {
-              Ref: 'VPC',
-            },
-          },
-        },
-        PublicNetworkAclEgress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '0.0.0.0/0',
-            NetworkAclId: {
-              Ref: 'PublicNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-        PublicNetworkAclIngress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '0.0.0.0/0',
-            NetworkAclId: {
-              Ref: 'PublicNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-        PublicSubnetNetworkAclAssociation1: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: 'PublicSubnet1',
-            NetworkAclId: 'PublicNetworkAcl',
-          },
-        },
-        PublicSubnetNetworkAclAssociation2: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: 'PublicSubnet2',
-            NetworkAclId: 'PublicNetworkAcl',
-          },
-        },
-        PublicSubnetNetworkAclAssociation3: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'PublicSubnet3',
-            },
-            NetworkAclId: {
-              Ref: 'PublicNetworkAcl',
-            },
-          },
-        },
-      };
-      const actual = plugin.buildPublicNetworkAcl(3);
-      expect(actual).toEqual(expected);
-    });
-  });
-
-  describe('#buildAppNetworkAcl', () => {
-    it('builds the application network ACL', () => {
-      const expected = {
-        AppNetworkAcl: {
-          Type: 'AWS::EC2::NetworkAcl',
-          Properties: {
-            Tags: [
-              {
-                Key: 'STAGE',
-                Value: 'dev',
-              },
-              {
-                Key: 'Name',
-                Value: {
-                  'Fn::Join': [
-                    '-',
-                    [
-                      {
-                        Ref: 'AWS::StackName',
-                      },
-                      'app',
-                    ],
-                  ],
-                },
-              },
-            ],
-            VpcId: {
-              Ref: 'VPC',
-            },
-          },
-        },
-        AppNetworkAclEgress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.8.0/22',
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-        AppNetworkAclIngress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.8.0/22',
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-        AppNetworkAclEgress101: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.24.0/22',
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 101,
-          },
-        },
-        AppNetworkAclIngress101: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.24.0/22',
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 101,
-          },
-        },
-        AppNetworkAclEgress102: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.40.0/22',
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 102,
-          },
-        },
-        AppNetworkAclIngress102: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.40.0/22',
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 102,
-          },
-        },
-        AppSubnetNetworkAclAssociation1: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'AppSubnet1',
-            },
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-          },
-        },
-        AppSubnetNetworkAclAssociation2: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'AppSubnet2',
-            },
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-          },
-        },
-        AppSubnetNetworkAclAssociation3: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'AppSubnet3',
-            },
-            NetworkAclId: {
-              Ref: 'AppNetworkAcl',
-            },
-          },
-        },
-      };
-      const publicSubnets = [
-        '10.0.8.0/22',
-        '10.0.24.0/22',
-        '10.0.40.0/22',
-      ];
-      const actual = plugin.buildAppNetworkAcl(publicSubnets);
-      expect(actual).toEqual(expected);
-    });
-  });
-
-  describe('#buildDBNetworkAcl', () => {
-    it('builds the database network ACL', () => {
-      const expected = {
-        DBNetworkAcl: {
-          Type: 'AWS::EC2::NetworkAcl',
-          Properties: {
-            Tags: [
-              {
-                Key: 'STAGE',
-                Value: 'dev',
-              },
-              {
-                Key: 'Name',
-                Value: {
-                  'Fn::Join': [
-                    '-',
-                    [
-                      {
-                        Ref: 'AWS::StackName',
-                      },
-                      'db',
-                    ],
-                  ],
-                },
-              },
-            ],
-            VpcId: {
-              Ref: 'VPC',
-            },
-          },
-        },
-        DBNetworkAclEgress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.0.0/21',
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-        DBNetworkAclIngress100: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.0.0/21',
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 100,
-          },
-        },
-        DBNetworkAclEgress101: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.16.0/21',
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 101,
-          },
-        },
-        DBNetworkAclIngress101: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.16.0/21',
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 101,
-          },
-        },
-        DBNetworkAclEgress102: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.32.0/21',
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-            Egress: true,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 102,
-          },
-        },
-        DBNetworkAclIngress102: {
-          Type: 'AWS::EC2::NetworkAclEntry',
-          Properties: {
-            CidrBlock: '10.0.32.0/21',
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-            Egress: false,
-            Protocol: -1,
-            RuleAction: 'allow',
-            RuleNumber: 102,
-          },
-        },
-        DBSubnetNetworkAclAssociation1: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'DBSubnet1',
-            },
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-          },
-        },
-        DBSubnetNetworkAclAssociation2: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'DBSubnet2',
-            },
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-          },
-        },
-        DBSubnetNetworkAclAssociation3: {
-          Type: 'AWS::EC2::SubnetNetworkAclAssociation',
-          Properties: {
-            SubnetId: {
-              Ref: 'DBSubnet3',
-            },
-            NetworkAclId: {
-              Ref: 'DBNetworkAcl',
-            },
-          },
-        },
-      };
-      const appSubnets = [
-        '10.0.0.0/21',
-        '10.0.16.0/21',
-        '10.0.32.0/21',
-      ];
-      const actual = plugin.buildDBNetworkAcl(appSubnets);
       expect(actual).toEqual(expected);
     });
   });
