+++ example/package.json
@@ -10,7 +10,7 @@
   },
   "main": "handler.js",
   "scripts": {
-    "deploy": "SLS_DEBUG=1 ./node_modules/.bin/sls deploy -v --force",
+    "deploy": "./node_modules/.bin/sls deploy -v --force",
     "remove": "./node_modules/.bin/sls remove -v",
     "test": "./node_modules/.bin/sls invoke -f s3"
   },
@@ -19,6 +19,6 @@
   },
   "devDependencies": {
     "serverless": "1.33.2",
-    "serverless-vpc-plugin": "smoketurner/serverless-vpc-plugin#jp-gh6"
+    "serverless-vpc-plugin": "smoketurner/serverless-vpc-plugin#master"
   }
 }
